requirements.txt
flask
Pillow
sentence_transformers
faiss-cpu
flask_cors
groq
